erp_production/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15
django-filter>=24.0
django-cors-headers>=4.4
celery>=5.4
requests>=2.32
Pillow>=11.0
python-decouple>=3.8
gunicorn>=23.0
whitenoise>=6.7
psycopg2-binary>=2.9
